+++ UsefulMap/UsefulMap/AboutApp/Model/WelcomeCard.swift
