+++ UsefulMap/UsefulMap/AboutApp/ViewModel/AboutAppViewModel.swift
