--- UsefulMap/UsefulMap/Extentions/GlobalProperties/File.swift
@@ -1,15 +1,0 @@
-//
-//  File.swift
-//  UsefulMap
-//
-//  Created by Dmitry Zasenko on 08.12.22.
-//
-
-import SwiftUI
-
-struct AppImages {
-    static let iconBar = UIImage(named: "barIcon")
-    static let iconRestaurant = UIImage(named: "restaurantIcon")
-    static let iconCafe = UIImage(named: "cafeIcon")
-    static let iconClub = UIImage(named: "clubIcon")
-}
