+++ UsefulMap/UsefulMap/GlobalProperties/AppImages.swift
@@ -1,0 +1,15 @@
+//
+//  AppImages.swift
+//  UsefulMap
+//
+//  Created by Dmitry Zasenko on 08.12.22.
+//
+
+import SwiftUI
+
+struct AppImages {
+    static let iconBar = UIImage(named: "barIcon")
+    static let iconRestaurant = UIImage(named: "restaurantIcon")
+    static let iconCafe = UIImage(named: "cafeIcon")
+    static let iconClub = UIImage(named: "clubIcon")
+}
