+++ UsefulMap/UsefulMap/GlobalProperties/SizesConstants.swift
@@ -1,0 +1,12 @@
+//
+//  SizesConstants.swift
+//  UsefulMap
+//
+//  Created by Eduard on 03.12.2022.
+//
+
+import Foundation
+
+struct SizesConstants {
+    static let logoSize: CGFloat = 60
+}
