+++ UsefulMap/UsefulMap/Map/Model/CustomAnnotation.swift
