--- UsefulMap/UsefulMap/Model/Cities.swift
@@ -1,16 +1,0 @@
-//
-//  Cities.swift
-//  UsefulMap
-//
-//  Created by Dmitry Zasenko on 29.11.22.
-//
-
-import Foundation
-
-typealias Cities = [City]
-
-struct City: Codable, Identifiable {
-    var id: Int = 0
-    var name: String = ""
-    var places: DecodedPlaces = []
-}
