--- UsefulMap/UsefulMap/Model/SizesConstants.swift
@@ -1,12 +1,0 @@
-//
-//  GlobalProperties.swift
-//  UsefulMap
-//
-//  Created by Eduard on 03.12.2022.
-//
-
-import Foundation
-
-struct GlobalProperties {
-    static let logoSize: CGFloat = 60
-}
