+++ UsefulMap/UsefulMap/Models/AdditionalMethods/ClearUserDefaults.swift
