--- UsefulMap/UsefulMap/View/Autentification/AutentificationView.swift
@@ -1,30 +1,0 @@
-//
-//  AutentificationView.swift
-//  UsefulMap
-//
-//  Created by Dmitry Zasenko on 01.12.22.
-//
-
-import SwiftUI
-
-struct AutentificationView: View {
-    
-    //MARK: - Properties
-    
-    let userViewModel = UserViewModel()
-    let networkManager = NetworkManager()
-    
-    //MARK: - Private properties
-    
-    @State private var isLocationViewOpen = false
-    
-    //MARK: - Body
-    
-    var body: some View {
-        if isLocationViewOpen {
-            LocationView( networkManager: networkManager, userViewModel: userViewModel)
-        } else {
-            LoginView(userViewModel: userViewModel, networkManager: networkManager, isLocationViewOpen: $isLocationViewOpen)
-        }
-    }
-}
